add manage user and all loan page for admin

+++ src/components/Dashboard/TableRows/ManageLoanDataRow.jsx
@@ -2,32 +2,44 @@ import axios from "axios";
 import React from "react";
 import toast from "react-hot-toast";
 import { Link } from "react-router";
+import Swal from "sweetalert2";
 
 const ManageLoanDataRow = ({ loan, refetch }) => {
-  const handleUpdate = () => {
-    console.log("update btn");
+  const handleDelete = (id) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "This loan will be permanently deleted.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#6b7280",
+      confirmButtonText: "Yes, Delete",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const res = await axios.delete(
+            `${import.meta.env.VITE_API_URL}/loan-delete/${id}`
+          );
 
-
-  };
-
-  const handleDelete = async (id) => {
-    try {
-      const res = await axios.delete(
-        `${import.meta.env.VITE_API_URL}/loan-delete/${id}`
-      );
-
-      if (res.data.deletedCount > 0) {
-        toast.success("Loan deleted successfully!");
-        refetch();
+          if (res.data.deletedCount > 0) {
+            Swal.fire({
+              title: "Deleted!",
+              text: "Loan has been successfully deleted.",
+              icon: "success",
+              confirmButtonColor: "#16a34a",
+            });
+            refetch();
+          }
+        } catch (err) {
+          console.error("Error:", err);
+          toast.error("Failed to delete loan. Try again.");
+        }
       }
-    } catch (err) {
-      console.error("Error:", err);
-      toast.error("Failed to delete loan. Try again.");
-    }
+    });
   };
 
   return (
-    <tr className="border-b border-gray-200">
+    <tr className="border-b">
       <td className="px-5 py-4 bg-white text-sm text-center">
         <img
           src={loan.image}
@@ -51,7 +63,7 @@ const ManageLoanDataRow = ({ loan, refetch }) => {
       <td className="px-5 py-4 bg-white text-sm text-center">
         <div className="flex justify-center items-center gap-2">
           <Link
-            to={`update-loan/${loan._id}`}
+            to={`/update-loan/${loan._id}`}
             className="px-3 py-1 bg-blue-600 text-white rounded hover:bg-blue-700"
           >
             Update
